% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{cellmatch}
\alias{cellmatch}
\title{cellmatch}
\format{
An object of class \code{data.frame} with 49625 rows and 11 columns.
}
\source{
\url{https://github.com/ZJUFanLab/scCATCH/tree/master/data}
}
\usage{
cellmatch
}
\description{
Marker genes of \code{'Human'} and \code{'Mouse'}.
}
\keyword{datasets}
