% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{createscCATCH}
\alias{createscCATCH}
\title{scCATCH object}
\usage{
createscCATCH(data, cluster)
}
\arguments{
\item{data}{A matrix or dgCMatrix containing normalized single-cell RNA-seq data, each column representing a cell, each row representing a gene. See \code{\link{demo_data}}.}

\item{cluster}{A character containing the cluster information for each cell. The length of it must be equal to the ncol of the data.}
}
\value{
scCATCH object
}
\description{
create scCATCH object using single-cell count data and cluster information.
}
