% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_geneinfo}
\alias{demo_geneinfo}
\title{Demo data of geneinfo}
\usage{
demo_geneinfo()
}
\description{
Demo data of geneinfo
}
\details{
\code{geneinfo} used in \code{\link{rev_gene}} must be a \code{data.frame} object with three columns, namely \code{'symbol'}, \code{'synonyms'}, \code{'species'}.
}
\examples{
geneinfo_demo <- demo_geneinfo()
}
