% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo_marker}
\alias{demo_marker}
\title{Demo data of markers}
\usage{
demo_marker()
}
\description{
Demo data of markers
}
\details{
\code{markers} used in \code{\link{findmarkergene}} must be a \code{data.frame} object with eleven columns.
}
\examples{
markers_demo <- demo_marker()
}
