% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{findcelltype,scCATCH-method}
\alias{findcelltype,scCATCH-method}
\title{Evidence-based score and annotation for each cluster}
\usage{
\S4method{findcelltype}{scCATCH}(object, verbose = TRUE)
}
\arguments{
\item{object}{scCATCH object generated from \code{\link{findmarkergene}}.}

\item{verbose}{Show progress messages.}
}
\value{
scCATCH object containing the results of predicted cell types for each cluster.
}
\description{
Evidence-based score and annotation for each cluster.
}
