% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmarkergenes.R
\name{findmarkergenes}
\alias{findmarkergenes}
\title{Find marker genes for each cluster}
\usage{
findmarkergenes(object, cell_min_pct = 0.25, logfc = 0.25,
  pvalue = 0.05)
}
\arguments{
\item{object}{Seurat object (>= 3.0.0) after log10 normalization and cluster analysis.}

\item{cell_min_pct}{Include the gene detected in at least this many cells in each cluster.}

\item{logfc}{Include the gene with at least this fold change of average gene expression compared to every other clusters.}

\item{pvalue}{Include the significantly highly expressed gene with this cutoff of p value from wilcox test compared to every other clusters.}
}
\value{
A data.frame containing cluster marker genes and the corresponding expressed cells percentage and average fold change for each cluster.
}
\description{
Identify marker genes for each cluster from a Seurat object (>= 3.0.0) after log10 normaliztion and cluser analysis. The marker gene in each cluster is identified according to its expression level compared to it in every other clusters. Only significantly highly expressed one in all pair-wise comparison of the cluster will be selected as a cluster marker gene.
}
\examples{
findmarkergenes(mouse_kidney_203_Seurat, 0.25, 0.25, 0.05)
}
