% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{geneinfo}
\format{
An object of class \code{data.frame} with 250934 rows and 3 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/gene}
}
\usage{
geneinfo
}
\description{
Gene symbols of \code{'Human'} and \code{'Mouse'} updated on Jan. 1, 2022 for revising genes.
}
\keyword{datasets}
