% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_kidney_203_Seurat.R
\docType{data}
\name{mouse_kidney_203_Seurat}
\alias{mouse_kidney_203_Seurat}
\title{mouse_kidney_203_Seurat}
\format{An object of class \code{Seurat} with 46959 rows and 203 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/?term=29089413}
}
\usage{
mouse_kidney_203_Seurat
}
\description{
A Seurat object containing mouse 203 kidney cells.
}
\keyword{datasets}
