% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{findcelltype}
\alias{findcelltype}
\title{Evidence-based score and annotation for each cluster}
\usage{
findcelltype(object)
}
\arguments{
\item{object}{scCATCH object generated from \code{\link{findmarkergene}}.}
}
\value{
scCATCH object containing the results of predicted cell types for each cluster.
}
\description{
Evidence-based score and annotation for each cluster.
}
