% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{scCATCH}
\alias{scCATCH}
\alias{scCATCH-class}
\title{Definition of 'scCATCH' class}
\description{
An S4 class containing the data, meta, and results of inferred cell types.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing normalized data. See \code{\link{demo_data}}.}

\item{\code{meta}}{A data frame containing the meta data.}

\item{\code{para}}{A list containing the parameters.}

\item{\code{markergene}}{A data frame containing the identified markers for each cluster.}

\item{\code{celltype}}{A data frame containing the cell types for each cluster.}

\item{\code{marker}}{A data frame containing the known markers. See \code{\link{demo_marker}}.}
}}

